requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.3
einops==0.8.1
fastapi==0.120.2
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
imageio==2.37.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.1
kornia-rs==0.1.9
lazy-loader==0.4
llvmlite==0.45.1
matplotlib==3.10.7
ml-dtypes==0.5.3
numba==0.62.1
numpy==2.1.3
onnx==1.19.1
onnx-ir==0.1.12
onnxruntime==1.23.2
onnxscript==0.5.4
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.0
pydantic==2.12.3
pydantic-core==2.41.4
pymatting==1.1.14
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
pytz==2025.2
referencing==0.37.0
regex==2025.10.22
rembg==2.0.67
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.20
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
